perf(backend): share one embedder process-wide; drop dead modules
embedding_url and services.recommendation_pipeline each kept their own
module-level _embedder and each constructed all-mpnet-base-v2. They run in the
same Flask process, so that was two copies of identical weights (~420 MB each)
resident for the process lifetime. Both now call one lazy accessor in
services/embedder.py.

Verified with SentenceTransformer stubbed: constructed once, and
embedding_url.get_embedder() is embedding_url.get_embedder() is
recommendation_pipeline.get_embedder(). Loading stays lazy — importing the
module costs nothing.

get_embedder stays bound in embedding_url's namespace, so the existing
monkeypatch in tests/test_embedding_url.py and the `from embedding_url import
get_embedder` in test_dpga_real_positives.py keep working unchanged.

Dead code removed, all confirmed to have no importers:
  - embedding_description.py — loaded a third SentenceTransformer (MiniLM)
  - gh_cleaner.py — referenced only by a commented-out line in models/app.py,
    removed here too
  - `from transformers import pipeline` in embedding_url.py — never used
  - the trailing `SDGs = ["SDG 1", "SDG 2", "SDG 3"]` list in sdg_constants.py,
    flagged as leftover in CLAUDE.md

96 tests pass.
